+++ gnp/src/main/java/controllers/SalesController.java
@@ -1,4 +1,4 @@
-package com.gnp.auth;
+package controllers;
 
 import java.math.*;
 import java.util.ArrayList;
@@ -8,6 +8,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+
+import models.Sales;
+import repos.SalesRepository;
 
 @RestController
 public class SalesController {
