+++ gnp/src/main/java/repos/UserRepository.java
@@ -1,6 +1,8 @@
-package com.gnp.auth;
+package repos;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import models.User;
 
 public interface UserRepository extends JpaRepository<User, Long> {
   User findByUsername(String username);
