+++ gnp/src/main/java/secuJwt/JWTAuthorizationFilter.java
@@ -1,4 +1,4 @@
-package com.gnp.auth;
+package secuJwt;
 
 import java.io.IOException;
 import java.util.ArrayList;
