+++ gnp/src/main/resources/application.properties
@@ -2,3 +2,6 @@ spring.jpa.hibernate.ddl-auto=update
 spring.datasource.url=jdbc:mysql://localhost:3306/gnp?createDatabaseIfNotExist=true
 spring.datasource.username=root
 spring.datasource.password=Password1!
+spring.jpa.properties.hibernate.dialect= org.hibernate.dialect.MySQL5InnoDBDialect
+gnp.app.jwtSecret= RestInPepperoni
+gnp.app.jwtExpirationMs= 432000000
